moved data types to Table, fixed removePlayer
